TheWorldOfTomorrow
The world of tomorrow is a business database created for the department of air traffic controlling the flying licenses, flying tests, employees, and etc. It has relational tables and has already gone under 4 normalization tests. Additionally, there are 10 queries at the bottom of the code looking for specific information that can be found in the database