Add files via upload
Minesweeper is a three class java project imitating the game minesweeper. The three classes are tile, the board, and the driver class. The tile holds information on what it displays. The board sets up where the bombs are and answer to commands of the driver class in regard to which tile the user has selected. The driver class models the basic rules of game, tying together the entire game.