LibraryModel
This library model is a four class java project imitating the functions of a library where a user can take books out and put them back in. The four classes are the book, the bookshelf, the person, and the driver class. The book contains basic information on itself such as its title, author, and etc. The bookshelf holds instances of books in each position. It also monitors what books are taken out and put back in. The person class holds information on who the person is. Lastly, the driver class connects the person class and bookshelf class in which allows the person to borrow and return the books.